// src/types/supabase.ts
export interface definitions {
  appointment: {
    /**
     * Format: bigint
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /** Format: uuid */
    user_id: string;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    created_at?: string;
    /** Format: text */
    name: string;
    /** Format: json */
    date: string;
    /** Format: json */
    requestor: string;
    /** Format: json */
    invitee: string;
    /** Format: ARRAY */
    invitee_reminder_periods: ReminderPeriod[];
    /** Format: ARRAY */
    requestor_reminder_periods: ReminderPeriod[];
    /** Format: text */
    timezone: TimezoneType;
  };
}